refactor(core): address code review findings
- Fix double-disposal bug in SteamProcessHelper
- Add validation for malformed shortcuts.vdf in BinaryKv
- Extract BackupManager from ShortcutsLibrary
- Extract ImportExportService from ShortcutsLibrary
- Extract WriteBackHandler from ShortcutsLibrary
- Replace batch state fields with ExportContext
- Remove unused GetGameActionDetails method
- Add theme-aware colors to UI dialogs
