+++ src/SteamShortcutsImporter/ExportContext.cs
@@ -1,0 +1,8 @@
+namespace SteamShortcutsImporter;
+
+internal class ExportContext
+{
+    public bool SkipAllRemaining { get; set; } = false;
+
+    public bool SteamCheckPerformed { get; set; } = false;
+}
